Many changes to many files
Important changes are: UIInvoker class allows safe interaction with
objects using UI controls from any thread, Added IRCServer and cleaned
up other code in the IRC namespace, Fixed broken UnitTests

+++ DesktopCS/Behaviors/AutoScrollBehavior.cs
@@ -22,7 +22,7 @@ namespace DesktopCS.Behaviors
 
         private void _scrollViewer_LayoutUpdated(object sender, EventArgs e)
         {
-            if (_scrollViewer.VerticalOffset == _scrollViewer.ScrollableHeight)
+            if ((int)_scrollViewer.VerticalOffset == (int)_scrollViewer.ScrollableHeight)
                 _scrollViewer.ScrollToBottom();
         }
